refactor: utils/buttons 생성, Button, ChoiceButton에 적용
